refactor(au-compose): expose composing and composition as bindables
